+++ packages/experimental/Avatar/src/JSAvatarTokens.ts
@@ -1,13 +1,13 @@
 import { Theme } from '@fluentui-react-native/framework';
 import { TokenSettings } from '@fluentui-react-native/use-styling';
-import { AvatarTokens } from '.';
-import { convertCoinColorFluent } from './Avatar.helpers';
+import { JSAvatarTokens } from '.';
+import { convertCoinColorFluent } from './JSAvatar.helpers';
 
-export const defaultAvatarTokens: TokenSettings<AvatarTokens, Theme> = () =>
+export const defaultJSAvatarTokens: TokenSettings<JSAvatarTokens, Theme> = () =>
   ({
     horizontalIconAlignment: 'end',
     verticalIconAlignment: 'end',
     color: 'white', // initials is always 'white', unless overriden by token
     iconStrokeColor: 'white', // icon stroke color is always 'white', unless overriden by token
     backgroundColor: convertCoinColorFluent('cornflower'),
-  } as AvatarTokens);
+  } as JSAvatarTokens);