+++ packages/experimental/Avatar/src/index.ts
@@ -1,2 +1,3 @@
 export * from './Avatar';
-export * from './Avatar.types';
+export * from './JSAvatar';
+export * from './JSAvatar.types';