+++ packages/experimental/Button/src/FAB/FABCore.tsx
@@ -1,13 +1,14 @@
 /** @jsx withSlots */
 import * as React from 'react';
 import { View } from 'react-native';
-import { buttonName, ButtonType, ButtonProps } from './Button.types';
+import { fabName, FABType } from './FAB.types';
 import { Text } from '@fluentui-react-native/experimental-text';
-import { stylingSettings, getDefaultSize } from './Button.styling';
+import { stylingSettings } from './FAB.styling';
 import { compose, mergeProps, withSlots, UseSlots } from '@fluentui-react-native/framework';
-import { useButton } from './useButton';
+import { useButton } from '../useButton';
 import { Icon } from '@fluentui-react-native/icon';
 import { createIconProps, IPressableState } from '@fluentui-react-native/interactive-hooks';
+import { ButtonCoreProps } from '../Button.types';
 
 /**
  * A function which determines if a set of styles should be applied to the compoent given the current state and props of the button.
@@ -17,35 +18,31 @@ import { createIconProps, IPressableState } from '@fluentui-react-native/interac
  * @param userProps The props that were passed into the button
  * @returns Whether the styles that are assigned to the layer should be applied to the button
  */
-export const buttonLookup = (layer: string, state: IPressableState, userProps: ButtonProps): boolean => {
-  return (
-    state[layer] ||
-    userProps[layer] ||
-    layer === userProps['appearance'] ||
-    layer === userProps['size'] ||
-    (!userProps['size'] && layer === getDefaultSize()) ||
-    (layer === 'hasContent' && userProps.content) ||
-    (layer === 'hasIcon' && userProps.icon)
-  );
+const buttonLookup = (layer: string, state: IPressableState, userProps: ButtonCoreProps): boolean => {
+  return state[layer] || userProps[layer] || (layer === 'hasContent' && userProps.content) || (layer === 'hasIcon' && userProps.icon);
 };
 
-export const Button = compose<ButtonType>({
-  displayName: buttonName,
+export const FAB = compose<FABType>({
+  displayName: fabName,
   ...stylingSettings,
   slots: {
     root: View,
     icon: Icon,
     content: Text,
   },
-  render: (userProps: ButtonProps, useSlots: UseSlots<ButtonType>) => {
-    const button = useButton(userProps);
+  render: (userProps: ButtonCoreProps, useSlots: UseSlots<FABType>) => {
+    const { icon, content, onClick, ...rest } = userProps;
     const iconProps = createIconProps(userProps.icon);
+
+    const button = useButton(rest);
+
     // grab the styled slots
     const Slots = useSlots(userProps, (layer) => buttonLookup(layer, button.state, userProps));
 
     // now return the handler for finishing render
-    return (final: ButtonProps, ...children: React.ReactNode[]) => {
-      const { icon, content, ...mergedProps } = mergeProps(button.props, final);
+    return (final: ButtonCoreProps, ...children: React.ReactNode[]) => {
+      const mergedProps = mergeProps(button.props, final);
+
       return (
         <Slots.root {...mergedProps}>
           {icon && <Slots.icon {...iconProps} />}
@@ -57,4 +54,4 @@ export const Button = compose<ButtonType>({
   },
 });
 
-export default Button;
+export default FAB;