+++ apps/macos/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/tester-macos
 
-This log was last generated on Wed, 17 Nov 2021 19:28:07 GMT and should not be manually modified.
+This log was last generated on Wed, 17 Nov 2021 21:07:06 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.1
+
+Wed, 17 Nov 2021 21:07:06 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.48.1 (sanajmi@microsoft.com)
 
 ## 0.0.1